feat: restructure weather module with new components and utility functions
